refactor job functions to new file
